requirements.txt
alembic==1.13.1
attrs==23.2.0
certifi==2024.8.30
cffi==1.16.0
charset-normalizer==3.3.2
greenlet==3.0.3
h11==0.14.0
idna==3.8
Mako==1.3.5
MarkupSafe==2.1.5
outcome==1.3.0.post0
psycopg2==2.9.9
pycparser==2.22
PySocks==1.7.1
python-decouple==3.8
requests==2.32.3
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.32
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
wsproto==1.2.0
anyio==4.4.0
httpcore==1.0.5
httpx==0.27.2
python-dotenv==1.0.1
python-telegram-bot==21.5
psycopg2-binary==2.9.9
